+++ components/CarCard.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import backendHandler from "../services/BackendHandler";
 import {CarCardPropsForReal} from "../screens/ScreenParams";
 import {getCarCardStylesheet} from "../services/CarCardStylesheet";
-import {playHapticFeedbackMultipleTimes} from "./HapticFeedback";
+import {playHapticFeedbackMultipleTimes} from "../services/HapticFeedback";
 
 function CarCard({params, navigation}: CarCardPropsForReal ){
     const car = params.car
