+++ services/HapticFeedback.ts
